UBSan fixes

+++ tests/UnitTests/TestInvDwtFrame.cc
@@ -287,7 +287,8 @@ class FRAME_IDWT : public ::testing::TestWithParam<fixture_param_t> {
         for (uint32_t c = 0; c < num_comp; c++) {
             for (uint32_t i = 0; i < lenght; i++) {
                 int16_t coeff = rnd->Rand16();
-                buffers_in_out_ref_32bit[c][i] = coeff << picture_header_dynamic.hdr_Fq;
+                /* UBSan fix: use LSHIFT to avoid UB when left-shifting negative int16_t coefficients. */
+                buffers_in_out_ref_32bit[c][i] = (int32_t)((uint32_t)(int32_t)coeff << picture_header_dynamic.hdr_Fq);
                 //Currently IDWT also shift-left by Fq internally to reduce memory bandwidth
                 buffers_in_mod_16bit[c][i] = coeff;
                 buffers_out_mod_32bit[c][i] = buffers_in_out_ref_32bit[c][i];
